refactor: wip normalize variables resource
