+++ ui/src/normalizers/localStorage/resources.ts
@@ -3,6 +3,7 @@ import {RemoteDataState, LocalStorage} from 'src/types'
 
 export const normalizeResources = (state: LocalStorage) => {
   return {
+    variables: state.resources.variables,
     orgs: normalizeOrgs(state.resources.orgs),
   }
 }
