+++ ui/src/store/configureStore.ts
@@ -74,6 +74,7 @@ export const rootReducer = combineReducers<ReducerState>({
     tasks: tasksReducer,
     telegrafs: telegrafsReducer,
     tokens: authsReducer,
+    variables: variablesReducer,
   }),
   routing: routerReducer,
   rules: rulesReducer,
@@ -83,7 +84,6 @@ export const rootReducer = combineReducers<ReducerState>({
   templates: templatesReducer,
   timeMachines: timeMachinesReducer,
   userSettings: userSettingsReducer,
-  variables: variablesReducer,
   variableEditor: variableEditorReducer,
   views: viewsReducer,
   VERSION: () => '',
