+++ ui/src/types/resources.ts
@@ -7,6 +7,7 @@ import {
   Telegraf,
   Scraper,
   TasksState,
+  VariablesState,
 } from 'src/types'
 
 export enum ResourceType {
@@ -52,4 +53,5 @@ export interface ResourceState {
   [ResourceType.Telegrafs]: TelegrafsState
   [ResourceType.Scrapers]: NormalizedState<Scraper>
   [ResourceType.Tasks]: TasksState
+  [ResourceType.Variables]: VariablesState
 }
