+++ ui/src/types/variables.ts
@@ -1,8 +1,18 @@
 import {Variable as GenVariable} from 'src/client'
 export {VariableProperties} from 'src/client'
-import {Label} from 'src/types'
+
+import {
+  VariableArgumentType,
+  QueryArguments,
+  MapArguments,
+  CSVArguments,
+  Label,
+  RemoteDataState,
+  NormalizedState,
+} from 'src/types'
 
 export interface Variable extends GenVariable {
+  status: RemoteDataState // Loading status of an individual variable
   labels: Label[]
 }
 
@@ -29,4 +39,25 @@ export interface VariableValuesByID {
 
 export interface ValueSelections {
   [variableID: string]: string
+}
+
+export interface VariablesState extends NormalizedState<Variable> {
+  values: {
+    // Different variable values can be selected in different
+    // "contexts"---different parts of the app like a particular dashboard, or
+    // the Data Explorer
+    [contextID: string]: {
+      status: RemoteDataState
+      order: string[] // IDs of variables
+      values: VariableValuesByID
+    }
+  }
+}
+
+export interface VariableEditorState {
+  name: string
+  selected: VariableArgumentType
+  argsQuery: QueryArguments
+  argsMap: MapArguments
+  argsConstant: CSVArguments
 }
