+++ ui/src/variables/selectors/index.tsx
@@ -15,6 +15,7 @@ import {
 import {VariableAssignment} from 'src/types/ast'
 import {
   AppState,
+  ResourceState,
   VariableArguments,
   VariableArgumentType,
   Variable,
@@ -23,19 +24,19 @@ import {
   ValueSelections,
 } from 'src/types'
 
-type VariablesState = AppState['variables']['variables']
-type ValuesState = AppState['variables']['values']['contextID']
+type VariablesState = ResourceState['variables']['byID']
+type ValuesState = ResourceState['variables']['values']['contextID']
 
 const extractVariablesListMemoized = memoizeOne(
   (variablesState: VariablesState): Variable[] => {
-    return Object.values(variablesState)
-      .filter(d => d.status === RemoteDataState.Done)
-      .map(d => d.variable)
+    return Object.values(variablesState).filter(
+      d => d.status === RemoteDataState.Done
+    )
   }
 )
 
 export const extractVariablesList = (state: AppState): Variable[] => {
-  return extractVariablesListMemoized(state.variables.variables)
+  return extractVariablesListMemoized(state.resources.variables.byID)
 }
 
 export const extractVariableEditorName = (state: AppState): string => {
@@ -100,10 +101,14 @@ export const getVariablesForDashboard = (
   state: AppState,
   dashboardID: string
 ): Variable[] => {
-  const variableIDs = get(state, `variables.values.${dashboardID}.order`, [])
+  const variableIDs = get(
+    state,
+    `resources.variables.values.${dashboardID}.order`,
+    []
+  )
 
   return getVariablesForDashboardMemoized(
-    state.variables.variables,
+    state.resources.variables.byID,
     variableIDs
   )
 }
@@ -113,18 +118,17 @@ export const getValuesForVariable = (
   variableID: string,
   contextID: string
 ): VariableValues => {
-  return get(state, `variables.values.${contextID}.values.${variableID}`)
+  return get(
+    state,
+    `resources.variables.values.${contextID}.values.${variableID}`
+  )
 }
 
 export const getTypeForVariable = (
   state: AppState,
   variableID: string
 ): VariableArguments['type'] => {
-  return get(
-    state,
-    `variables.variables.${variableID}.variable.arguments.type`,
-    ''
-  )
+  return get(state, `resources.variables.byID.${variableID}.arguments.type`, '')
 }
 
 type ArgumentValues = {[key: string]: string} | string[]
@@ -135,7 +139,7 @@ export const getArgumentValuesForVariable = (
 ): ArgumentValues => {
   return get(
     state,
-    `variables.variables.${variableID}.variable.arguments.values`,
+    `resources.variables.byID.${variableID}.arguments.values`,
     {}
   )
 }
@@ -145,7 +149,7 @@ export const getValueSelections = (
   contextID: string
 ): ValueSelections => {
   const contextValues: VariableValuesByID =
-    get(state, `variables.values.${contextID}.values`) || {}
+    get(state, `resources.variables.values.${contextID}.values`) || {}
 
   const selections: ValueSelections = Object.keys(contextValues).reduce(
     (acc, k) => {
@@ -175,7 +179,7 @@ const getVariableAssignmentsMemoized = memoizeOne(
     const result: VariableAssignment[] = Object.entries(
       valuesState.values
     ).reduce((acc, [variableID, values]) => {
-      const variableName = get(variablesState, [variableID, 'variable', 'name'])
+      const variableName = get(variablesState, [variableID, 'name'])
 
       if (!variableName || !values || !values.selectedValue) {
         return acc
@@ -193,8 +197,8 @@ export const getVariableAssignments = (
   contextID: string
 ): VariableAssignment[] =>
   getVariableAssignmentsMemoized(
-    state.variables.values[contextID],
-    state.variables.variables
+    state.resources.variables.values[contextID],
+    state.resources.variables.byID
   )
 
 export const getTimeMachineValues = (
@@ -204,7 +208,7 @@ export const getTimeMachineValues = (
   const activeTimeMachineID = state.timeMachines.activeTimeMachineID
   const values = get(
     state,
-    `variables.values.${activeTimeMachineID}.values.${variableID}`
+    `resources.variables.values.${activeTimeMachineID}.values.${variableID}`
   )
 
   return values
@@ -216,7 +220,7 @@ export const getTimeMachineValuesStatus = (
   const activeTimeMachineID = state.timeMachines.activeTimeMachineID
   const valuesStatus = get(
     state,
-    `variables.values.${activeTimeMachineID}.status`
+    `resources.variables.values.${activeTimeMachineID}.status`
   )
 
   return valuesStatus
@@ -225,18 +229,18 @@ export const getTimeMachineValuesStatus = (
 export const getDashboardVariablesStatus = (
   state: AppState
 ): RemoteDataState => {
-  return get(state, 'variables.status')
+  return get(state, 'resources.variables.status')
 }
 
 export const getDashboardValuesStatus = (
   state: AppState,
   dashboardID: string
 ): RemoteDataState => {
-  return get(state, `variables.values.${dashboardID}.status`)
+  return get(state, `resources.variables.values.${dashboardID}.status`)
 }
 
 export const getVariable = (state: AppState, variableID: string): Variable => {
-  return get(state, `variables.variables.${variableID}.variable`)
+  return get(state, `resources.variables.byID.${variableID}`)
 }
 
 export const getHydratedVariables = (
@@ -244,12 +248,12 @@ export const getHydratedVariables = (
   contextID: string
 ): Variable[] => {
   const hydratedVariableIDs: string[] = Object.keys(
-    get(state, `variables.values.${contextID}.values`, {})
+    get(state, `resources.variables.values.${contextID}.values`, {})
   )
 
-  const hydratedVariables = Object.values(state.variables.variables)
-    .map(d => d.variable)
-    .filter(v => hydratedVariableIDs.includes(v.id))
+  const hydratedVariables = Object.values(
+    state.resources.variables.byID
+  ).filter(v => hydratedVariableIDs.includes(v.id))
 
   return hydratedVariables
 }
