+++ ui/src/variables/utils/hydrateVars.test.ts
@@ -6,7 +6,7 @@ import {hydrateVars} from 'src/variables/utils/hydrateVars'
 import {createMapVariable, createVariable} from 'src/variables/mocks'
 
 // Types
-import {Variable, CancellationError} from 'src/types'
+import {Variable, CancellationError, RemoteDataState} from 'src/types'
 
 class FakeFetcher implements ValueFetcher {
   responses = {}
@@ -190,6 +190,7 @@ describe('hydrate vars', () => {
           k: 'v',
         },
       },
+      status: RemoteDataState.NotStarted,
     }
 
     const vars = [a, b]
@@ -213,7 +214,7 @@ describe('hydrate vars', () => {
     }).promise
 
     // Basic test for now, we would need an icky mock to assert that the
-    // approriate substitution is actually taking place
+    // appropriate substitution is actually taking place
     expect(actual.a.error).toBeFalsy()
     expect(actual.b.error).toBeFalsy()
   })
@@ -230,6 +231,7 @@ describe('hydrate vars', () => {
         type: 'constant',
         values: ['v1', 'v2'],
       },
+      status: RemoteDataState.NotStarted,
     }
 
     const vars = [a, b]
